+++ src/Android/TimeBase.hpp
@@ -17,10 +17,10 @@ namespace Android {
 		static TimeBase* REAL_TIME;
 
 	public:
-		virtual int64_t getTime(const ThreadData* paramThreadData, Call::CallList* callList) = 0;
-		virtual int64_t getElapsedInclusiveTime(const MethodData* paramMethodData) = 0;
-		virtual int64_t getElapsedExclusiveTime(const MethodData* paramMethodData) = 0;
-		virtual int64_t getElapsedInclusiveTime(const ProfileData* paramProfileData) = 0;
+		virtual uint32_t getTime(const ThreadData* paramThreadData, Call::CallList* callList) = 0;
+		virtual uint32_t getElapsedInclusiveTime(const MethodData* paramMethodData) = 0;
+		virtual uint32_t getElapsedExclusiveTime(const MethodData* paramMethodData) = 0;
+		virtual uint32_t getElapsedInclusiveTime(const ProfileData* paramProfileData) = 0;
 	};
 
 	class CpuTimeBase final
@@ -28,10 +28,10 @@ namespace Android {
 	{
 	public:
 		static CpuTimeBase TIME;
-		virtual int64_t getTime(const ThreadData* threadData, Call::CallList* callList) override;
-		virtual int64_t getElapsedInclusiveTime(const MethodData* methodData) override;
-		virtual int64_t getElapsedExclusiveTime(const MethodData* methodData) override;
-		virtual int64_t getElapsedInclusiveTime(const ProfileData* profileData) override;
+		virtual uint32_t getTime(const ThreadData* threadData, Call::CallList* callList) override;
+		virtual uint32_t getElapsedInclusiveTime(const MethodData* methodData) override;
+		virtual uint32_t getElapsedExclusiveTime(const MethodData* methodData) override;
+		virtual uint32_t getElapsedInclusiveTime(const ProfileData* profileData) override;
 	};
 
 	class RealTimeBase final
@@ -39,9 +39,9 @@ namespace Android {
 	{
 	public:
 		static RealTimeBase TIME;
-		virtual int64_t getTime(const ThreadData* threadData, Call::CallList* callList) override;
-		virtual int64_t getElapsedInclusiveTime(const MethodData* methodData) override;
-		virtual int64_t getElapsedExclusiveTime(const MethodData* methodData) override;
-		virtual int64_t getElapsedInclusiveTime(const ProfileData* profileData) override;
+		virtual uint32_t getTime(const ThreadData* threadData, Call::CallList* callList) override;
+		virtual uint32_t getElapsedInclusiveTime(const MethodData* methodData) override;
+		virtual uint32_t getElapsedExclusiveTime(const MethodData* methodData) override;
+		virtual uint32_t getElapsedInclusiveTime(const ProfileData* profileData) override;
 	};
 };
